app: pure refactor/reorganize to separate *Queue and APIJob objects
We are going to be adding quite a few more API jobs (even without
considering any additional functionality that might be added to the
application). Let's separate the logic out a bit so that it's easier
to move around the codebase.
